agentictx/backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6

# Validation
pydantic==2.10.4
pydantic-settings==2.7.0

# AI
anthropic==0.42.0

# File processing
python-docx==1.1.2
PyMuPDF==1.24.14
pillow==11.1.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1

# Dev / test
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
